src/robot/chat/chat/requirements.txt
ackermann-msgs==2.0.2
action-msgs==1.2.1
action-tutorials-interfaces==0.20.3
action-tutorials-py==0.20.3
actionlib-msgs==4.2.3
aiofiles==23.2.1
aiohttp==3.9.0
aiosignal==1.3.1
altair==5.2.0
ament-cmake-test==1.3.6
ament-copyright==0.12.9
ament-cppcheck==0.12.9
ament-cpplint==0.12.9
ament-flake8==0.12.9
ament-index-python==1.4.0
ament-lint==0.12.9
ament-lint-cmake==0.12.9
ament-package==0.14.0
ament-pep257==0.12.9
ament-uncrustify==0.12.9
ament-xmllint==0.12.9
angles==1.15.0
annotated-types==0.6.0
anyascii==0.3.2
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
bond==3.0.2
builtin-interfaces==1.2.1
cachetools==5.3.2
cartographer-ros-msgs==2.0.9000
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.18
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
composition-interfaces==1.2.1
control-msgs==4.4.0
controller==2023.1.1
controller-manager==2.35.0
controller-manager-msgs==2.35.0
cv-bridge==3.2.1
dataclasses-json==0.6.2
demo-nodes-py==0.20.3
Deprecated==1.2.14
diagnostic-msgs==4.2.3
diagnostic-updater==3.1.2
distro==1.8.0
domain-coordinator==0.10.0
dwb-msgs==1.1.12
example-interfaces==0.9.3
examples-rclpy-executors==0.15.1
examples-rclpy-minimal-action-client==0.15.1
examples-rclpy-minimal-action-server==0.15.1
examples-rclpy-minimal-client==0.15.1
examples-rclpy-minimal-publisher==0.15.1
examples-rclpy-minimal-service==0.15.1
examples-rclpy-minimal-subscriber==0.15.1
exceptiongroup==1.2.0
faiss-cpu==1.7.4
faiss-gpu==1.7.2
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.10.0
gazebo-msgs==3.7.0
generate-parameter-library-py==0.3.6
geometry-msgs==4.2.3
google-auth==2.23.4
googleapis-common-protos==1.61.0
gradio==4.7.1
gradio_client==0.7.0
greenlet==3.0.1
grpcio==1.59.3
gTTS==2.4.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.4
image-geometry==3.2.1
importlib-metadata==6.8.0
importlib-resources==6.1.1
inflect==7.0.0
interactive-markers==2.3.2
Jinja2==3.1.2
joblib==1.3.2
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kubernetes==28.1.0
langchain==0.0.339
langsmith==0.0.66
laser-geometry==2.4.0
launch==1.0.4
launch-ros==0.19.6
launch-testing==1.0.4
launch-testing-ros==0.19.6
launch-xml==1.0.4
launch-yaml==1.0.4
lifecycle-msgs==1.2.1
llvmlite==0.41.1
logging-demo==0.20.3
map-msgs==2.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
message-filters==4.3.3
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nav-2d-msgs==1.1.12
nav-msgs==4.2.3
nav2-common==1.1.12
nav2-msgs==1.1.12
nav2-simple-commander==1.0.0
networkx==3.2.1
nltk==3.8.1
numba==0.58.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnx==1.15.0
onnxruntime==1.16.3
openai==1.3.5
openai-whisper==20231117
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
orjson==3.9.10
osrf-pycommon==2.0.2
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pcl-msgs==1.0.0
pendulum-msgs==0.20.3
Pillow==10.1.0
posthog==3.0.2
protobuf==4.25.1
pulsar-client==3.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.17.2
pypdf==3.17.1
PyPika==0.48.9
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-qt-binding==1.1.2
PyYAML==6.0.1
qt-dotgraph==2.2.2
qt-gui==2.2.2
qt-gui-cpp==2.2.2
qt-gui-py-common==2.2.2
quality-of-service-demo-py==0.20.3
rcl-interfaces==1.2.1
rclpy==3.3.11
rcutils==5.1.4
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
resource-retriever==3.1.1
rich==13.7.0
rmw-dds-common==1.6.0
ros2action==0.18.7
ros2bag==0.15.8
ros2cli==0.18.7
ros2component==0.18.7
ros2controlcli==2.35.0
ros2doctor==0.18.7
ros2interface==0.18.7
ros2launch==0.19.6
ros2lifecycle==0.18.7
ros2multicast==0.18.7
ros2node==0.18.7
ros2param==0.18.7
ros2pkg==0.18.7
ros2run==0.18.7
ros2service==0.18.7
ros2topic==0.18.7
rosbag2-interfaces==0.15.8
rosbag2-py==0.15.8
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rpds-py==0.13.2
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.4
rqt-bag-plugins==1.1.4
rqt-console==2.0.2
rqt-graph==1.3.0
rqt-gui==1.1.5
rqt-gui-py==1.1.5
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.5
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.4
scripts==3.7.0
semantic-version==2.10.0
sensor-msgs==4.2.3
sensor-msgs-py==4.2.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
shape-msgs==4.2.3
shellingham==1.5.4
slam-toolbox==2.6.6
smclib==3.0.2
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
SpeechRecognition==3.10.0
SQLAlchemy==2.0.23
sros2==0.10.4
starlette==0.27.0
statistics-msgs==1.2.1
std-msgs==4.2.3
std-srvs==4.2.3
stereo-msgs==4.2.3
sympy==1.12
teleop-twist-keyboard==2.3.2
tenacity==8.2.3
tf-transformations==1.0.0
tf2-geometry-msgs==0.25.5
tf2-kdl==0.25.5
tf2-msgs==0.25.5
tf2-py==0.25.5
tf2-ros-py==0.25.5
tf2-tools==0.25.5
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.15.0
tomlkit==0.12.0
toolz==0.12.0
topic-monitor==0.20.3
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
trajectory-msgs==4.2.3
transformers==4.35.2
transforms3d==0.4.1
triton==2.1.0
ttstokenizer==1.0.0
turtle-tf2-py==0.3.6
turtlebot3-example==2.1.5
turtlebot3-msgs==2.2.3
turtlebot3-teleop==2.1.5
turtlesim==1.4.2
txtai==6.2.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
unique-identifier-msgs==2.2.1
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
vehicle==2023.1.1
vision-msgs==4.1.0
visualization-msgs==4.2.3
watchfiles==0.21.0
webots-ros2==2023.1.1
webots-ros2-driver==2023.1.1
webots-ros2-epuck==2023.1.1
webots-ros2-importer==2023.1.1
webots-ros2-mavic==2023.1.1
webots-ros2-msgs==2023.1.1
webots-ros2-tesla==2023.1.1
webots-ros2-tiago==2023.1.1
webots-ros2-turtlebot==2023.1.1
webots-ros2-universal-robot==2023.1.1
websocket-client==1.6.4
websockets==11.0.3
wrapt==1.16.0
xacro==2.0.8
yarl==1.9.3
zipp==3.17.0
